thing: interactor: Fix int validation
When JSON objects are decoded into a structure that contains generic
values, integer numbers are converted to float64 ones. Thus, it causes
a bug when validating data operations against the thing's schema. To
solve it, this patch validates if the number is really an integer when
asserting if it comes as float64.

+++ pkg/thing/interactors/publish_data_test.go
@@ -98,7 +98,7 @@ var publishDataUseCases = []PublishDataTestCase{
 		"data sensorId doesn't match with thing's schema",
 		"authorization-token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 1, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 1, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
@@ -113,7 +113,7 @@ var publishDataUseCases = []PublishDataTestCase{
 		"error publishing message in connector exchange",
 		"authorization-token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 0, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 0, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
@@ -128,7 +128,7 @@ var publishDataUseCases = []PublishDataTestCase{
 		"message successfully sent to connector exchange",
 		"authorization-token",
 		"thing-id",
-		[]entities.Data{entities.Data{SensorID: 0, Value: 5}},
+		[]entities.Data{entities.Data{SensorID: 0, Value: float64(5)}},
 		&mocks.FakeLogger{},
 		&mocks.FakeThingProxy{Thing: &entities.Thing{
 			ID:     "thing-id",
